backend: Move find_kernel_in_sysroot to backend/kernel.py
Small refactor moving find_kernel_in_sysroot() from "backend/common.py"
to "backend/kernel.py".

Related-to: TCB-762
